requirements.txt
altgraph==0.15
beautifulsoup4==4.6.0
blinker==1.3
bs4==0.0.1
chardet==2.3.0
configobj==5.0.6
crontab==0.22.1
cryptography==1.2.3
Django==2.0.6
djangorestframework==3.8.2
future==0.16.0
huepy==0.9.8.1
idna==2.0
instabot==0.4.1
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9

macholib==1.9
MarkupSafe==0.23
oauthlib==1.0.3
peewee==3.5.0
pefile==2017.11.5
Pillow==5.1.0
prettytable==0.7.2
py2exe==0.9.2.2
pyasn1==0.1.9
pycurl==7.43.0

PyInstaller==3.3.1
PyJWT==1.3.0
pyserial==3.0.1

python-crontab==2.2.8
python-dateutil==2.7.2


pytz==2018.4
PyYAML==3.11
requests==2.9.1
requests-toolbelt==0.8.0
schedule==0.5.0
six==1.10.0
ssh-import-id==5.5
tqdm==4.23.4


urllib3==1.13.1
